benchmarks/gpu/stress/requirements.txt
# GPU Stress Test Dependencies
# Core dependencies
numpy>=1.21.0
psutil>=5.8.0
pyyaml>=6.0

# PyTorch with CUDA support
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# GPU monitoring
nvidia-ml-py3>=7.352.0

# Optional: Additional GPU optimization
# apex  # Install manually if needed 